base: use ptrdiff_t for array and string sizes and some other things.

+++ src/input/input.h
@@ -26,26 +26,26 @@ public:
     const String& GetName() const { return m_name; }
 
     /** Gets the index of the corresponding key, needed to call GetKey */
-    int GetKeyIndex(const char* name) const { return GetItemIndex(name, m_keynames); }
+    ptrdiff_t GetKeyIndex(const char* name) const { return GetItemIndex(name, m_keynames); }
     /** Gets the index of the corresponding axis, needed to call GetAxis */
-    int GetAxisIndex(const char* name) const { return GetItemIndex(name, m_axisnames); }
+    ptrdiff_t GetAxisIndex(const char* name) const { return GetItemIndex(name, m_axisnames); }
     /** Gets the index of the corresponding cursor, needed to call GetCursor */
-    int GetCursorIndex(const char* name) const { return GetItemIndex(name, m_cursornames); }
+    ptrdiff_t GetCursorIndex(const char* name) const { return GetItemIndex(name, m_cursornames); }
 
     /** Gets the current state of the given key, true being pressed and false being released */
-    bool GetKey(int index) const { return m_keys[index]; }
+    bool GetKey(ptrdiff_t index) const { return m_keys[index]; }
     /** Gets the current value of the given axis. Devices should cap this value between -1 and 1 as much as possible, through it is not guaranteed */
-    float GetAxis(int index) const { return m_axis[index].m1 * m_axis[index].m2; }
+    float GetAxis(ptrdiff_t index) const { return m_axis[index].m1 * m_axis[index].m2; }
     /** Gets the current value of the given cursor, 0,0 being the bottom-left corner and 1,1 being the top-right corner */
-    vec2 GetCursor(int index) const { return m_cursors[index].m1; }
+    vec2 GetCursor(ptrdiff_t index) const { return m_cursors[index].m1; }
     /** Gets the coordinate of the pixel the cursor is currently over, 0,0 being the bottom-left corner. */
-    ivec2 GetCursorPixel(int index) const { return m_cursors[index].m2; }
+    ivec2 GetCursorPixel(ptrdiff_t index) const { return m_cursors[index].m2; }
 
 
     /** Sets a per-device-axis sensitivity factor. The value returned by the operating system will be multiplied by this value before being returned by GetAxis */
-    void SetAxisSensitivity(int index, float sensitivity) { m_axis[index].m2 = sensitivity; }
+    void SetAxisSensitivity(ptrdiff_t index, float sensitivity) { m_axis[index].m2 = sensitivity; }
     /** Gets the per-device-axis sensitivity factor. The value returned by the operating system will be multiplied by this value before being returned by GetAxis */
-    float GetAxisSensitivity(int index) const { return m_axis[index].m2; }
+    float GetAxisSensitivity(ptrdiff_t index) const { return m_axis[index].m2; }
 
     /** Gets a list of the name of all available keys in this device */
     const array<String>& GetAllKeys() const { return m_keynames; }
@@ -86,7 +86,7 @@ protected:
 
     ~InputDevice()
     {
-        for (int i = 0; i < devices.Count(); ++i)
+        for (ptrdiff_t i = 0; i < devices.Count(); ++i)
         {
             if (devices[i] == this)
             {
@@ -100,9 +100,9 @@ private:
     static array<InputDevice*> devices;
 
     template <typename... T>
-    int GetItemIndex(const char* name, const array<String, T...>& array) const
+    ptrdiff_t GetItemIndex(const char* name, const array<String, T...>& array) const
     {
-        for (int i = 0; i < array.Count(); ++i)
+        for (ptrdiff_t i = 0; i < array.Count(); ++i)
         {
             if (array[i] == name)
                 return i;
@@ -112,7 +112,7 @@ private:
 
     static InputDevice* GetDevice(const char* name)
     {
-        for (int i = 0; i < devices.Count(); ++i)
+        for (ptrdiff_t i = 0; i < devices.Count(); ++i)
         {
             if (devices[i]->m_name == name)
                 return devices[i];
